Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.fontVariant.settings-expected.txt

2d.text.drawing.style.fontVariant.settings

Testing basic functionalities of fontKerning for canvas


FAIL Testing basic functionalities of fontKerning for canvas assert_equals: ctx.fontVariantCaps === "Normal" (got [undefined], expected Normal[string]) expected (string) "Normal" but got (undefined) undefined
